Crate bb8

A full-featured connection pool, designed for asynchronous connections (using tokio). Originally based on r2d2.

Opening a new database connection every time one is needed is both inefficient and can lead to resource exhaustion under high traffic conditions. A connection pool maintains a set of open connections to a database, handing them out for repeated use.

bb8 is agnostic to the connection type it is managing. Implementors of the ManageConnection trait provide the database-specific logic to create and check the health of connections.

Example

Using an imaginary “foodb” database.

#[tokio::main]
async fn main() {
    let manager = bb8_foodb::FooConnectionManager::new("localhost:1234");
    let pool = bb8::Pool::builder().build(manager).await.unwrap();

    for _ in 0..20 {
        let pool = pool.clone();
        tokio::spawn(async move {
            let conn = pool.get().await.unwrap();
            // use the connection
            // it will be returned to the pool when it falls out of scope.
        });
    }
}

Structs

Builder

A builder for a connection pool.

NopErrorSink

An ErrorSink implementation that does nothing.

Pool

A generic connection pool.

PooledConnection

A smart pointer wrapping a connection.

State

Information about the state of a Pool.

Enums

QueueStrategy

bb8’s queue strategy when getting pool resources

RunError

bb8’s error type.

Traits

CustomizeConnection

A trait which provides functionality to initialize a connection

ErrorSink

A trait to receive errors generated by connection management that aren’t tied to any particular caller.

ManageConnection

A trait which provides connection-specific functionality.